requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.2
psycopg2==2.9.1
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
soupsieve==2.2.1
sqlparse==0.4.2
tenacity==8.0.1
urllib3==1.26.7
whitenoise==5.3.0
